+++ docs/changelog.rst
@@ -6,19 +6,18 @@
 Release Notes
 *************
 
-Version 0.1.5 (TBD)
+Version 0.2.0 (2017-11-22)
 -------------------
 
 **New Features**:
 
-* Added ``dagman_progress`` command line tool for displaying a progress bar for Dagman jobs. (See `PR #45 <https://github.com/jrbourbeau/pycondor/pull/45>`_)
+* Added ``dagman_progress`` command line tool for displaying a progress bar for Dagman jobs. (See `PR #45 <https://github.com/jrbourbeau/pycondor/pull/45>`_ and `PR #52 <https://github.com/jrbourbeau/pycondor/pull/52>`_)
 * Added environment variable option for setting submit, error, log, and output directories. (See `PR #50 <https://github.com/jrbourbeau/pycondor/pull/50>`_)
 
 **Bug Fixes**:
 
 * Fixed bug where the queue parameter for a Job was not written to the job submit file when the Job was built by a Dagman. (See `PR #42 <https://github.com/jrbourbeau/pycondor/pull/42>`_)
 * Fixed bug that caused a filename mismatch between a ``Job`` submit file and the error/log/output files when a named argument is added to a Job, and the Job is built with ``fancyname=True``. (See `PR #48 <https://github.com/jrbourbeau/pycondor/pull/48>`_)
-* Fixed issue that caused ``dagman_progress`` to round up the percentage of jobs that were marked as done. (See `PR #52 <https://github.com/jrbourbeau/pycondor/pull/52>`_)
 * Fixed the Dagman submit file build procedure to include the name of Job named arguments in the Dagman node name (See `PR #53 <https://github.com/jrbourbeau/pycondor/pull/53>`_)